now using size_t instead of int

+++ evstyunichev.konstantin/S1/list_actions.hpp
@@ -3,7 +3,7 @@
 #include <iostream>
 #include <list>
 
-using pair_t = std::pair< std::string, std::list< int > >;
+using pair_t = std::pair< std::string, std::list< size_t> >;
 
 namespace evstyunichev
 {
